Update ng-true/false-value (html only)
This is not a change to the directive itself.  However, in the html here: http://ionicframework.com/docs/api/directive/ionToggle/
 you have the following code snippet:


 <ion-toggle toggle-class="toggle-assertive"
                    ng-model="emailNotification"
                    ng-true-value="Subscribed"
                    ng-false-value="Unubscribed">
          Newsletter
        </ion-toggle>

This doesn't work because there are no single quotes on 'Subscribed' and 'Unsubscribed', it should be like this:
ng-true-value="'Subscribed'"

Not a big deal, just had me scratching my head for a few minutes and might be nice to update and keep the docs pretty!  
By the way, I love Ionic....Thanks!

Note that I added one space to the end of the file simply to enable myself to submit this, you can just reject this pull request